# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for monitoring Keycloak instances, managing alerts, and security operations",
        "title": "Keycloak Monitoring Tool API",
        "termsOfService": "https://defensepoint.com/terms",
        "contact": {
            "name": "DefensePoint Support",
            "url": "https://defensepoint.com/support",
            "email": "support@defensepoint.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth/callback": {
            "get": {
                "description": "Handles callback from OAuth2 provider after authentication",
                "tags": [
                    "Auth"
                ],
                "summary": "OAuth2 callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF state parameter",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to home page"
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/login/oauth2": {
            "get": {
                "description": "Redirects to OAuth2 provider for authentication",
                "tags": [
                    "Auth"
                ],
                "summary": "Initiate OAuth2 login",
                "responses": {
                    "302": {
                        "description": "Redirect to OAuth2 provider"
                    },
                    "403": {
                        "description": "OAuth2 disabled",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login/simple": {
            "post": {
                "description": "Authenticate using simple username/password credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login with username and password",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto_requests.SimpleLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "user": {
                                    "$ref": "#/definitions/auth.UserInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Simple auth disabled",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "description": "Clears the user session and logs out",
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "302": {
                        "description": "Redirect to home page"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/password/change": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change the password for a user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Password change request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto_requests.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Simple auth disabled",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refreshes an expired access token using the refresh token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "expiry": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token refresh not available",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/userinfo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns information about the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_chi.UserDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse": {
            "description": "Standard error response with optional validation details",
            "type": "object",
            "properties": {
                "details": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "github_com_DefensePoint_keycloak-monitoring_internal_http_dto_requests.SimpleLogin": {
            "description": "Request body for simple authentication",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 1,
                    "example": "mypassword"
                },
                "username": {
                    "type": "string",
                    "minLength": 1,
                    "example": "johndoe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}